cyclonev22/requirements.txt
    # Cyclone v2 Project Requirements

    # --- Core & Scheduling ---
    python-dotenv
    schedule
    requests # Used for CryptoNews API and potentially others
    pandas>=1.3.0 # Ensure pandas version supports needed features
    numpy<2.0
    pytz # For timezone handling

    # --- Database ---
    sqlalchemy>=1.4.0 # Use modern SQLAlchemy
    # psycopg2-binary # Commented out - Not needed for Azure SQL
    pyodbc # Uncommented - Needed for Azure SQL/MS SQL Server

    # --- APIs ---
    python-binance # For Binance API
    pandas-ta # For Technical Indicators
    # cryptonews-api # Removed - No specific library found, using 'requests'
    praw # For Reddit API
    tweepy>=4.0.0 # For Twitter API v2

    # --- ML & Sentiment ---
    transformers[torch] # Or transformers[tensorflow]
    torch # Or tensorflow
    scikit-learn
    lightgbm # Or xgboost
    joblib # For saving/loading models

    # --- Dashboard ---
    dash>=2.0.0
    plotly
    dash-bootstrap-components # For better dashboard layout/styling

    # --- Utilities ---
    # Add any other utility libraries needed
    